+++ test/functional/kernel/tile-variants/tests/test-kernel-tile-LocalArray2D.hpp
@@ -1,17 +1,17 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-// Copyright (c) 2016-21, Lawrence Livermore National Security, LLC
+// Copyright (c) 2016-22, Lawrence Livermore National Security, LLC
 // and RAJA project contributors. See the RAJA/LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-#ifndef __TEST_KERNEL_TILE_FIXED2D_HPP__
-#define __TEST_KERNEL_TILE_FIXED2D_HPP__
+#ifndef __TEST_KERNEL_TILE_LOCALARRAY2D_HPP__
+#define __TEST_KERNEL_TILE_LOCALARRAY2D_HPP__
 
 #include <numeric>
 
 template <typename INDEX_TYPE, typename DATA_TYPE, typename WORKING_RES, typename EXEC_POLICY>
-void KernelTileFixed2DTestImpl(const int rows, const int cols)
+void KernelTileLocalArray2DTestImpl(const int rows, const int cols)
 {
   // This test emulates matrix transposition with tiling.
 
@@ -48,6 +48,10 @@ void KernelTileFixed2DTestImpl(const int rows, const int cols)
   RAJA::View<DATA_TYPE, RAJA::Layout<2>> WorkTView( work_array_t, cols, rows );
   RAJA::View<DATA_TYPE, RAJA::Layout<2>> CheckTView( check_array_t, cols, rows );
 
+  // initialize local array (shared mem)
+  using TILE_MEM = RAJA::LocalArray<DATA_TYPE, RAJA::Perm<0,1>, RAJA::SizeList<tile_dim_x, tile_dim_y>>;
+  TILE_MEM Tile_Array;
+
   // initialize arrays
   std::iota( test_array, test_array + array_length, 1 );
   std::iota( test_array_t, test_array_t + array_length, 1 );
@@ -68,10 +72,15 @@ void KernelTileFixed2DTestImpl(const int rows, const int cols)
   RAJA::TypedRangeSegment<INDEX_TYPE> rowrange( 0, rows );
   RAJA::TypedRangeSegment<INDEX_TYPE> colrange( 0, cols );
 
-  RAJA::kernel<EXEC_POLICY> ( RAJA::make_tuple( colrange, rowrange ),
-    [=] RAJA_HOST_DEVICE ( INDEX_TYPE cc, INDEX_TYPE rr ) {
-      WorkTView( cc, rr ) = WorkView( rr, cc );
-  });
+  RAJA::kernel_param<EXEC_POLICY> ( RAJA::make_tuple( colrange, rowrange ), RAJA::make_tuple( (INDEX_TYPE)0, (INDEX_TYPE)0, Tile_Array ),
+    [=] RAJA_HOST_DEVICE ( INDEX_TYPE cc, INDEX_TYPE rr, INDEX_TYPE tx, INDEX_TYPE ty, TILE_MEM &Tile_Array ) {
+      Tile_Array( ty, tx ) = WorkView( rr, cc );
+    },
+
+    [=] RAJA_HOST_DEVICE ( INDEX_TYPE cc, INDEX_TYPE rr, INDEX_TYPE tx, INDEX_TYPE ty, TILE_MEM &Tile_Array ) {
+      WorkTView( cc, rr ) = Tile_Array( ty, tx );
+    }
+  );
 
   work_res.memcpy( check_array_t, work_array_t, sizeof(DATA_TYPE) * array_length );
 
@@ -97,25 +106,25 @@ void KernelTileFixed2DTestImpl(const int rows, const int cols)
 }
 
 
-TYPED_TEST_SUITE_P(KernelTileFixed2DTest);
+TYPED_TEST_SUITE_P(KernelTileLocalArray2DTest);
 template <typename T>
-class KernelTileFixed2DTest : public ::testing::Test
+class KernelTileLocalArray2DTest : public ::testing::Test
 {
 };
 
-TYPED_TEST_P(KernelTileFixed2DTest, TileFixed2DKernel)
+TYPED_TEST_P(KernelTileLocalArray2DTest, TileLocalArray2DKernel)
 {
   using INDEX_TYPE  = typename camp::at<TypeParam, camp::num<0>>::type;
   using DATA_TYPE  = typename camp::at<TypeParam, camp::num<1>>::type;
   using WORKING_RES = typename camp::at<TypeParam, camp::num<2>>::type;
   using EXEC_POLICY = typename camp::at<TypeParam, camp::num<3>>::type;
 
-  KernelTileFixed2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(10, 10);
-  KernelTileFixed2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(151, 111);
-  KernelTileFixed2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(362, 362);
+  KernelTileLocalArray2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(10, 10);
+  KernelTileLocalArray2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(151, 111);
+  KernelTileLocalArray2DTestImpl<INDEX_TYPE, DATA_TYPE, WORKING_RES, EXEC_POLICY>(362, 362);
 }
 
-REGISTER_TYPED_TEST_SUITE_P(KernelTileFixed2DTest,
-                            TileFixed2DKernel);
+REGISTER_TYPED_TEST_SUITE_P(KernelTileLocalArray2DTest,
+                            TileLocalArray2DKernel);
 
-#endif  // __TEST_KERNEL_TILE_FIXED2D_HPP__
+#endif  // __TEST_KERNEL_TILE_LOCALARRAY2D_HPP__